backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pymupdf
requests
numpy
